Update Provider Configuration
Breaking changes to provider configuration:
- Token becomes optional
- Organization no longer deprecated
- `individual` and `anonymous` removed
- `owner` inferred from `organization`

This also refactors the provider configuration testing into a new 
`provider_utils` file alongside the original test file. The aim was to 
make this suite of tests more readable.
